Attempt to fix deployment bug
There is an issue with the 3rd party API access headers so I had to create a proxy for the API to bypass this so it can work correctly. However, the next js server components would not work with this new API url call. Therefore I had to change the code to a client component and implement useState and useEffect instead. I believe this should fix the issue and will display the data correctly after deployment.