Start creating the individual property page
Once a client has made a search, the client can click on any property and it will send him to the individual property page. Once that click happens, the property ID is set in the URL and once navigation is complete, the API call gets made with the ID taken from the params and displays the data correctly.